Unlock account and retry transfer if error occurs
In ZkSync, when a new account is going to move funds
for the first time, it needs to be "unlocked". So,
the first time an operation is attempted, it is going
to fail because of being locked.

This was done this way to avoid checking if the account
is unlocked every time. Unlocking the account only
has to be done once.

Signed-off-by: Rodimiro Cerrato E <rodyce@gmail.com>