Implement and test deposit, transfer and withdrawal ops (#6)
* Add tests for deposit, transfer and withdraw ops

* Unlock account and retry transfer if error occurs
In ZkSync, when a new account is going to move funds
for the first time, it needs to be "unlocked". So,
the first time an operation is attempted, it is going
to fail because of being locked.
This was done this way to avoid checking if the account
is unlocked every time. Unlocking the account only
has to be done once.

* Fix amount to send

* [zksync] Defer using signing wallet until needed
This improves UX.

* Separate unit tests in different suites

Signed-off-by: Rodimiro Cerrato E <rodyce@gmail.com>