Automate publishing to NPM registry checking pkg version
Publishing to NPM depends on the version number in the
local package.json vs published version in the NPM
public registry.

The only condition to block merge to the target branch
is that the version specified in the package.json file
is less than the one published in the NPM registry.
In any other case, the merge will NOT be blocked.
However, publish to NPM will ONLY be performed if the
version in the package.json is greater than the latest
version in the NPM registry or if no version is
published there.

In case of NPM failure, merge will still be possible,
but no publish will be performed.

Signed-off-by: Rodimiro Cerrato E <rodyce@gmail.com>